requirements.txt
requests>=2.32.0
python-dotenv>=1.0.0
pytz>=2024.1
bcrypt>=4.0.0
